Add implementation of BrokerApi for RabbitMQ
This change add an implementation of the BrokerApi from the
Gerrit events-broker lib. This make it possible to use
RabbitMQ as the message-broker in a multi-site setup.

The BrokerApi implementation can make use of the default
configuration values in $site_path/data/@PLUGIN@.config but
specific configurations needs to be set in
$site_path/data/@PLUGIN@/broker/broker.config

To make it possible to only use the BrokerApi implementation
you now need to explicitly put a config file in the site
directory if you want to use the original publisher
functionality.

Change-Id: I7806bee5a10e6a0724427ff2c1dd065c6af1514c