Add implementation of BrokerApi for RabbitMQ
This change add an implementation of the BrokerApi from the
Gerrit events-broker lib. This make it possible to use
RabbitMQ as the message-broker in a multi-site setup.

The BrokerApi implementation can make use of the default
configuration values in $site_path/data/@PLUGIN@.config but
specific configurations needs to be set in
$site_path/data/@PLUGIN@/broker/broker.config

Added a new section called General with the flags:

* publishAllGerritEvents: This will make the plugin publish gerrit
  stream events to a configured exchange. This was the original purpose
  of this plugin and is therefore enabled by default.
* enableBrokerApi: Enable the RabbitMQ Broker API. This is the new
  functionality and is disabled by default.

Change-Id: I7806bee5a10e6a0724427ff2c1dd065c6af1514c