Catch exceptions that consumers throw
As exceptions that are thrown by consumers are considered to
originate from the application, they cause a closure of the channel
that the consumers use. Neither channel nor consumer are recreated.
As consumers may still be able to process other events, we should
not close the channels due to exceptions. Instead warn, log the
exception and continue.

Change-Id: If4e85412af324568ef6ab5162e692529473298da