Make code-review -2 and +1 sticky
Often a CR+1 is given with a comment that needs to be solved. Making CR+1
sticky lets the first reviewer skip the round-trip of another CR+1 when
the comment is solved.

Since the definition of the Code-Review label on this repo diverges from
the default definition of the Code-Review label that is defined in
All-Projects we need to copy the full label definition into this repo.

Change-Id: Ia95545859162f3daa97f0f8f7e9fbe04b2bd9cab