Add support for consumer prefetch
If no consumer prefetch is set it becomes unlimited which
means that the rabbitmq server will send as many events as
it can, this can cause the client to be overwhelmed and make
it slow to respond. When the client respond to slowly the
server may decide to delete the consumer. With a limited
consumer prefetch we can prevent this.

Change-Id: I599aa024202597c468bd3f792ec0931f40a495f9