Add support for RabbitMQ streams
With streams you will be able to use the replay
functionality. To keep track of the current message an
offset is used. The offset is periodically saved/stored and
when replayAll is called the consumer start to consume from the
beginning of the stream.
Every stream.windowSize we store
(current offset - stream.windowSize) as the offset to make
sure we replay missed messages when we restart a multi-site
instance.

Change-Id: I51a49533248ce94d1e6831071b033869c86db6d8